rename utils/usage_metadata.ts to utils/metadata.ts
