+++ libs/providers/langchain-google/src/utils/metadata.ts
